finalbackend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
constantly==23.10.4
cryptography==43.0.1
cymem==2.0.8
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.1.0
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.9
incremental==24.7.2
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
thinc==8.3.2
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
websocket==0.2.1
wrapt==1.16.0
yarl==1.11.1
zope.event==5.0
zope.interface==7.0.3
